* Fixed countFrames(). * Changed timecodes to use a long for frame offset rather than an int

+++ core/core-video/src/main/java/org/openimaj/video/Video.java
@@ -130,7 +130,7 @@ public abstract class Video<T extends Image<?,T>> implements Iterable<T>
 	 * 
 	 * @param newFrame the new index
 	 */
-	public synchronized void setCurrentFrameIndex( int newFrame ) 
+	public synchronized void setCurrentFrameIndex( long newFrame ) 
 	{
 		// We're already at the frame?
 		if( this.currentFrame == newFrame ) return;
